+++ jigna/js/build.js
@@ -54,6 +54,7 @@ concatenate({
         'app/client.js',
         'app/async_client.js',
         'app/proxy_factory.js',
+        'app/async_proxy_factory.js',
         'app/proxy.js',
         'app/subarray.js',
         'app/list_proxy.js',
@@ -88,6 +89,7 @@ concatenate({
         'app/client.js',
         'app/async_client.js',
         'app/proxy_factory.js',
+        'app/async_proxy_factory.js',
         'app/proxy.js',
         'app/subarray.js',
         'app/list_proxy.js',
