+++ jigna/js/src/app/client.js
@@ -10,7 +10,7 @@ jigna.Client.prototype.initialize = function() {
 
     // Private protocol.
     this._id_to_proxy_map = {};
-    this._proxy_factory   = new jigna.ProxyFactory(this);
+    this._proxy_factory   = this._create_proxy_factory();
 
     // Add all of the models being edited
     jigna.add_listener(
@@ -226,6 +226,10 @@ jigna.Client.prototype._add_models = function(context) {
     jigna.ready.resolve();
 
     return models;
+};
+
+jigna.Client.prototype._create_proxy_factory = function() {
+    return new jigna.ProxyFactory(this);
 };
 
 jigna.Client.prototype._create_proxy = function(type, obj, info) {
