+++ jigna/js/src/app/proxy_factory.js
@@ -2,11 +2,6 @@
 // ProxyFactory
 ///////////////////////////////////////////////////////////////////////////////
 
-jigna._SavedData = function(data) {
-    // Used internally to save marshaled data to unmarshal later.
-    this.data = data;
-}
-
 jigna.ProxyFactory = function(client) {
     // Private protocol.
     this._client = client;
@@ -65,17 +60,7 @@ jigna.ProxyFactory.prototype._add_instance_attribute = function(proxy, attribute
         var value = this.__cache__[attribute_name];
         if (value === undefined) {
             value = this.__client__.get_attribute(this, attribute_name);
-            if (value === undefined) {
-                var info = this.__info__;
-                if (info && (info.attribute_values !== undefined)) {
-                    var index = info.attribute_names.indexOf(attribute_name);
-                    value = this.__client__._unmarshal(
-                        info.attribute_values[index]
-                    );
-                }
-            } else {
-                this.__cache__[attribute_name] = value;
-            }
+            this.__cache__[attribute_name] = value;
         }
 
         return value;
@@ -219,46 +204,17 @@ jigna.ProxyFactory.prototype._delete_dict_keys = function(proxy) {
 };
 
 jigna.ProxyFactory.prototype._populate_dict_proxy = function(proxy, info) {
-    var index, key;
-    var values = info.values;
+    var index;
 
-    if (values !== undefined) {
-        for (key in values.new_types) {
-            this._create_instance_constructor(values.new_types[key]);
-        }
-    }
-
-    for (index=0; index < info.keys.length; index++) {
-        key = info.keys[index];
-        this._add_item_attribute(proxy, key);
-        if (values !== undefined) {
-            proxy.__cache__[key] = new jigna._SavedData(values.data[index]);
-        }
+    for (index in info.keys) {
+        this._add_item_attribute(proxy, info.keys[index]);
     }
 };
 
 jigna.ProxyFactory.prototype._update_dict_proxy = function(proxy, info) {
-    var removed = info.removed;
-
-    if (removed !== undefined) {
-        // The async case has an additional removed key.
-        var cache = proxy.__cache__;
-        var key;
-
-        // Add the keys in the added.
-        this._populate_dict_proxy(proxy, info.added);
-
-        for (var index=0; index < removed.length; index++) {
-            key = removed[index];
-            delete cache[key];
-            delete proxy[key];
-        }
-    } else {
-        // The sync case.
-        proxy.__cache__ = {};
-        this._delete_dict_keys(proxy);
-        this._populate_dict_proxy(proxy, info);
-    }
+    proxy.__cache__ = {}
+    this._delete_dict_keys(proxy);
+    this._populate_dict_proxy(proxy, info);
 };
 
 // List proxy creation /////////////////////////////////////////////////////////
@@ -266,6 +222,7 @@ jigna.ProxyFactory.prototype._update_dict_proxy = function(proxy, info) {
 jigna.ProxyFactory.prototype._create_list_proxy = function(id, info) {
     var proxy = new jigna.ListProxy('list', id, this._client);
     this._populate_list_proxy(proxy, info);
+
     return proxy;
 };
 
@@ -280,60 +237,25 @@ jigna.ProxyFactory.prototype._delete_list_items = function(proxy) {
 jigna.ProxyFactory.prototype._populate_list_proxy = function(proxy, info) {
     /* Populate the items in a list proxy. */
 
-    for (var key in info.new_types) {
-        this._create_instance_constructor(info.new_types[key]);
-    }
-
-    var data = info.data;
     for (var index=0; index < info.length; index++) {
         this._add_item_attribute(proxy, index);
-        if (data !== undefined) {
-            proxy.__cache__[index] = new jigna._SavedData(data[index]);
-        }
     }
 
     return proxy;
 };
 
 jigna.ProxyFactory.prototype._update_list_proxy = function(proxy, info) {
-    /* Update the given proxy. */
-    if (info.added !== undefined) {
-        // For the async case
-        // Register any new types.
-        for (var key in info.added.new_types) {
-            this._create_instance_constructor(info.added.new_types[key]);
-        }
+    /* Update the given proxy.
+     *
+     * This removes all previous items and then repopulates the proxy with
+     * items that reflect the (possibly) new length.
+     */
+    this._delete_list_items(proxy);
+    this._populate_list_proxy(proxy, info);
 
-        var splice_args = [info.index, info.removed].concat(
-            info.added.data.map(function(x) {return new jigna._SavedData(x);})
-        );
-
-        var extra = splice_args.length - 2 - splice_args[1];
-        var cache = proxy.__cache__;
-        var end = cache.length;
-        if (extra < 0) {
-            for (var index=end; index > (end+extra) ; index--) {
-                delete proxy[index-1];
-            }
-        } else {
-            for (var index=0; index < extra; index++){
-                this._add_item_attribute(proxy, end+index);
-            }
-        }
-        cache.splice.apply(cache, splice_args);
-    } else {
-
-        /* This is the sync case.
-         * This removes all previous items and then repopulates the proxy with
-         * items that reflect the (possibly) new length.
-         */
-        this._delete_list_items(proxy);
-        this._populate_list_proxy(proxy, info);
-
-        // Get rid of any cached items (items we have already requested from
-        // the server-side.
-        proxy.__cache__ = []
-    }
+    // Get rid of any cached items (items we have already requested from the
+    // server-side.
+    proxy.__cache__ = []
 };
 
 // Common for list and dict proxies ////////////////////////////////////////////
@@ -346,9 +268,6 @@ jigna.ProxyFactory.prototype._add_item_attribute = function(proxy, index){
         var value = this.__cache__[index];
         if (value === undefined) {
             value = this.__client__.get_attribute(this, index);
-            this.__cache__[index] = value;
-        } else if (value instanceof jigna._SavedData) {
-            value = this.__client__._unmarshal(value.data);
             this.__cache__[index] = value;
         }
