+++ jigna/server.py
@@ -1,7 +1,5 @@
 #
-# Enthought product code
-#
-# (C) Copyright 2013 Enthought, Inc., Austin, TX
+# (C) Copyright 2013-2016 Enthought, Inc., Austin, TX
 # All right reserved.
 #
 """ Abstract base classes for the Jigna server and bridge. """
@@ -15,8 +13,8 @@ import traceback
 
 # Enthought library.
 from traits.api import (
-    Any, Bool, Dict, Event, HasTraits, Instance, Property, Str,
-    TraitDictEvent, TraitListEvent
+    Any, Dict, Event, HasTraits, Instance, Property, Str, TraitDictEvent,
+    TraitListEvent
 )
 
 # Logging.
@@ -38,10 +36,6 @@ class Server(HasTraits):
     """ Server that serves a Jigna view. """
 
     #### 'Server' protocol ####################################################
-
-    #: Use async communication, this is useful when the client only
-    # communicates via websockets.
-    async = Bool(False)
 
     #: Base url for serving content.
     base_url = Property(Str, depends_on='_base_url')
@@ -195,6 +189,7 @@ class Server(HasTraits):
 
     def get_instance_attribute(self, request):
         """ Get the value of an instance attribute. """
+
         obj            = self._id_to_object_map[request['id']]
         attribute_name = request['attribute_name']
 
@@ -287,30 +282,10 @@ class Server(HasTraits):
 
         return attribute_names
 
-    def _get_attribute_values(self, obj, attribute_names):
-        """ Get the values of all 'public' attributes on an object.
-
-        Return a list of strings.
-
-        """
-        def _getattr(obj, name):
-            v = getattr(obj, name)
-            if isinstance(v, list):
-                v = []
-            elif isinstance(v, dict):
-                v = {}
-            return v
-
-        return [self._marshal(_getattr(obj, name)) for name in attribute_names]
-
-
     def _get_dict_info(self, obj):
         """ Get a description of a dict. """
-        if self.async:
-            values = self._get_list_info(list(obj.values()))
-            return dict(keys=list(obj.keys()), values=values)
-        else:
-            return dict(keys=obj.keys())
+
+        return dict(keys=obj.keys())
 
     def _get_event_names(self, obj):
         """ Get the names of all the attributes on an object.
@@ -341,19 +316,13 @@ class Server(HasTraits):
         # need to include the full info for it (its attributes, events and
         # methods etc)...
         if type_name not in self._visited_type_names:
-            self._visited_type_names.add(type_name)
-            attribute_names = self._get_attribute_names(obj)
             info = dict(
-                type_name        = type_name,
-                attribute_names  = attribute_names,
-                event_names      = self._get_event_names(obj),
-                method_names     = self._get_public_method_names(type(obj))
+                type_name       = type_name,
+                attribute_names = self._get_attribute_names(obj),
+                event_names     = self._get_event_names(obj),
+                method_names    = self._get_public_method_names(type(obj))
             )
-            if self.async:
-                attribute_values = self._get_attribute_values(
-                    obj, attribute_names
-                )
-                info['attribute_values'] = attribute_values
+            self._visited_type_names.add(type_name)
 
         # ... for subsequent calls, we only need to send the type name as the
         # the client will have already built a prototype based on the previous
@@ -365,16 +334,8 @@ class Server(HasTraits):
 
     def _get_list_info(self, obj):
         """ Get a description of a list. """
-        if self.async:
-            data = self._marshal_all(obj)
-            new_types = {}
-            for x in data:
-                if x['type'] == 'instance' and len(x['info']) > 1:
-                    new_types[x['info']['type_name']] = x['info']
 
-            return dict(length=len(obj), data=data, new_types=new_types)
-        else:
-            return dict(length=len(obj))
+        return dict(length=len(obj))
 
     def _get_public_method_names(self, cls):
         """ Get the names of all public methods on a class.
@@ -480,10 +441,6 @@ class Server(HasTraits):
 
     def _send_object_changed_event(self, obj, trait_name, old, new):
         """ Send an object changed event. """
-        if self.async:
-            return self._send_object_changed_event_async(
-                obj, trait_name, old, new
-            )
 
         if trait_name.startswith('_'):
             return
@@ -507,58 +464,6 @@ class Server(HasTraits):
             # a type/value pair which we need on the client side to determine
             # what (if any) proxy we need to create.
             data = self._marshal(new),
-
-            # fixme: This is how we currently detect an 'xxx_items' event on
-            # the JS side.
-            items_event = items_event
-        )
-
-        self.send_event(event)
-
-        return
-
-    def _send_object_changed_event_async(self, obj, trait_name, old, new):
-        """ Send an object changed event. """
-
-        if trait_name.startswith('_'):
-            return
-
-        if isinstance(new, TraitListEvent):
-            trait_name  = trait_name[:-len('_items')]
-            value = id(getattr(obj, trait_name))
-            info = dict(
-                index=new.index, removed=len(new.removed),
-                added=self._get_list_info(new.added)
-            )
-            data = dict(type='list', value=value, info=info)
-            items_event = True
-
-        elif isinstance(new, TraitDictEvent):
-            trait_name  = trait_name[:-len('_items')]
-            trait = getattr(obj, trait_name)
-            value = id(trait)
-            added, removed = new.added, list(new.removed.keys())
-            for key in new.changed:
-                added[key] = trait[key]
-            info = dict(removed=removed, added=self._get_dict_info(added))
-            data = dict(type='dict', value=value, info=info)
-            items_event = True
-
-        else:
-            # fixme: intent is non-scalar or maybe container?
-            if hasattr(new, '__dict__') or isinstance(new, (dict, list)):
-                self._register_object(new)
-
-            data = self._marshal(new)
-            items_event = False
-
-        event = dict(
-            obj  = str(id(obj)),
-            name = trait_name,
-            # fixme: This smells a bit, but marshalling the new value gives us
-            # a type/value pair which we need on the client side to determine
-            # what (if any) proxy we need to create.
-            data = data,
 
             # fixme: This is how we currently detect an 'xxx_items' event on
             # the JS side.
